adjusting the default configuration of jest that the folder "_shared" does not get mapped into "app/shared"
The issue is that the module report-parameter has multiple components (custom input fields) which are used in the different report components. They have been moved into a submodule "_shared". The default jest configuration maps every module named "shared" to an absolute path which meant that the unit test was not able to find the code. The regex has been therefore adjusted only to map the global shared folder and not the _shared folder