Reintroduce birthday validator in the testform component
The validator for comparing the birthdate to the testing date stopped working (~year ago). Most likely because of an update of angular or oblique. The issue was that the value for the birthdate field was observed but the value obtained in the validator was NULL - essentially because the validator was executed before the field was able to update itself.

The necessary steps are:
- replacing the awfully hacky way how subform have been implemented (PersonalDataComponent)
- adding additional guards in the date validator for preventing errors
- adding observer on the birthdate field to ensure the validator gets triggered if any of the fields gets changed